requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bleach==6.2.0
cachetools==5.5.2
certifi==2023.7.22
charset-normalizer==3.4.2
dnspython==2.7.0
docstring_parser==0.16
google-api-core==2.24.2
google-auth==2.40.1
google-cloud-aiplatform==1.93.0
google-cloud-bigquery==3.32.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.15.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
kaggle==1.7.4.5
numpy==2.2.6
packaging==25.0
pandas==2.2.3
proto-plus==1.26.1
protobuf==6.31.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
pymongo==4.13.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
pytz==2025.2
requests==2.32.3
rsa==4.9.1
setuptools==80.7.1
shapely==2.1.0
six==1.17.0
sniffio==1.3.1
text-unidecode==1.3
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
webencodings==0.5.1
websockets==15.0.1
